Polished brainvision_wrangling file
Added more documentation and changed how segment files are saved and
handled. Removed the experiment crosscheck function due to possible
issues with making it for general use, giving this responsability to
each of the users of this code. A script file called
segment_data_with_tests.py is being developed containing an example.